SDP-1170 [pr-comments] Add `BeginTx` and `Request` to testutils package
